총알 데미지 설정(충돌 시 적용)
총알 데미지 설정
충돌 시 데미지 만큼 체력 감소

+++ 졸작 프로젝트/zombie/03. External/Include/Engine/MonsterScript.h
@@ -18,7 +18,7 @@ struct MonsterStatus
 	bool PlayerInRange = false;
 	bool PlayerInAttackRange = false;
 	bool isAttack = false;
-	int hp = 100;
+	float hp = 100;
 };
